# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Playbasis API",
        "description": "API Explorer",
        "termsOfService": "",
        "contact": {
            "email": "info@playbasis.com"
        },
        "version": "1.0"
    },
    "host": "api.pbapp.net",
    "basePath": "/",
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/Email/send": {
            "post": {
                "tags": [
                    "Communication"
                ],
                "description": "Send email to a player",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Access token returned by Playbasis Authentication",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "subject",
                        "in": "query",
                        "description": "Email subject",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "description": "Email message | message or template_id required",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "template_id",
                        "in": "query",
                        "description": "Template ID | message or template_id required",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/Email/goods": {
            "post": {
                "tags": [
                    "Communication"
                ],
                "description": "Send email to a player",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Access token returned by Playbasis Authentication",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "ref_id",
                        "in": "query",
                        "description": "Reference transaction ID for redemption",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "subject",
                        "in": "query",
                        "description": "Email subject",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "description": "Email message | message or template_id required",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "template_id",
                        "in": "query",
                        "description": "Template ID | message or template_id required",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/Email/recent": {
            "get": {
                "tags": [
                    "Communication"
                ],
                "description": "List recent emails sent to a player",
                "parameters": [
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "since",
                        "in": "query",
                        "description": "DateTime formatted string",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/Email/template/{template_id}": {
            "get": {
                "tags": [
                    "Communication"
                ],
                "description": "Get processed email template.",
                "parameters": [
                    {
                        "name": "template_id",
                        "in": "path",
                        "description": "Template message ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "apiKey": {
            "type": "apiKey",
            "name": "api_key",
            "in": "query"
        },
        "token": {
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    },
    "security": [
        {
            "apiKey": [],
            "token": []
        }
    ]
}